pyproject.toml
[project]
name = "tradyon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "feedparser>=6.0.12",
    "google-generativeai>=0.8.5",
    "investpy>=1.0.8",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langchain-google-genai>=2.0.10",
    "langchain-tavily>=0.2.12",
    "lxml>=6.0.2",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "streamlit>=1.50.0",
    "tavily-python>=0.7.12",
    "yfinance>=0.2.66",
]
